% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundie-package.r, R/boundie.r
\docType{package}
\name{boundie}
\alias{boundie}
\alias{boundie-package}
\alias{boundie}
\title{Bounded differential expression using count data}
\usage{
boundie(formula, data = NULL, weights = NULL, bounds = NULL,
  control = NULL)
}
\description{
Bounded differential expression using count data

Bounded coefficient differential expression
}
