% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit}
\alias{fit}
\title{Fit parameters using the optim function}
\usage{
fit(x, y, weights = rep(1, ncol(x)), offset = 0, control = list())
}
\arguments{
\item{x}{design matrix of dimension ‘n * p’}

\item{y}{vector of observations of length ‘n’}

\item{weights}{vector of weights of length ‘n’}

\item{offset}{vector of coefficient offsets of length ‘n’}

\item{control}{list of control paramters, e.g. ‘maxit’}
}
\value{
list with the following components:
  ‘par’ - best values for paramters found
  ‘value’ - negative log-likelihood at end of iterations
  ‘convergence’ - 0 if converged, other values indicate iteration limit or warning
  ‘message’ - character string with additional information
}
\description{
Fit parameters using the optim function
}
