% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.r
\name{gradient}
\alias{gradient}
\title{Gradient function for negative binomial regression}
\usage{
gradient(par, weights, offset)
}
\arguments{
\item{par}{Numeric vector of coefficient values}
}
\value{
Direction of steepest descent (the gradient)
}
\description{
The following variables need to be available to the function in order to
be able to call `optim` using this as a gradient function:
}
\details{
x - Regression coefficients matrix
beta - Regressors, such that beta * x = y
y - Observed count values
}
