% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.r
\name{lrt}
\alias{lrt}
\title{Likelihood ratio test for model fits}
\usage{
lrt(x, y, reduced, weights = rep(1, nrow(x)), offset = 0,
  control = list(), lower = -Inf, upper = Inf)
}
\arguments{
\item{x}{design matrix of dimension ‘n * p’}

\item{y}{vector of observations of length ‘n’}

\item{reduced}{character vector of parameters used in the reduced model}

\item{weights}{vector of weights of length ‘n’}

\item{offset}{vector of coefficient offsets of length ‘n’}

\item{control}{list of control paramters, e.g. ‘maxit’}

\item{lower}{vector of coefficient lower bounds}

\item{upper}{vector of coefficient upper bounds}
}
\value{
a data.frame with terms, their fitted values, and LRT result
}
\description{
Likelihood ratio test for model fits
}
