% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.r
\name{wald}
\alias{wald}
\title{Wald test for model fits}
\usage{
wald(x, y, weights = rep(1, nrow(x)), offset = 0, control = list(),
  lower = -Inf, upper = Inf, verbose = FALSE)
}
\arguments{
\item{x}{design matrix of dimension ‘n * p’}

\item{y}{vector of observations of length ‘n’}

\item{weights}{vector of weights of length ‘n’}

\item{offset}{vector of coefficient offsets of length ‘n’}

\item{control}{list of control paramters, e.g. ‘maxit’}

\item{lower}{vector of coefficient lower bounds}

\item{upper}{vector of coefficient upper bounds}

\item{verbose}{print debug messages}
}
\value{
a data.frame with terms, their fitted values, and wald statistic
}
\description{
Wald test for model fits
}
