% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundie-package.r, R/boundie.r
\docType{package}
\name{boundie}
\alias{boundie}
\alias{boundie-package}
\alias{boundie}
\title{Bounded differential expression using count data}
\usage{
boundie(x, design, weights = rep(1, ncol(x)), offset = rep(0, ncol(x)),
  lower = -Inf, upper = Inf, control = list())
}
\arguments{
\item{x}{a gene expression matrix (genes x samples)}

\item{design}{the model to fit for each gene, e.g. ‘~ condition‘}

\item{weights}{weight vector for all coefficients}

\item{offset}{vector of coefficient offsets of length ‘n’}

\item{lower}{vector of coefficient lower bounds}

\item{upper}{vector of coefficient upper bounds}

\item{control}{named list of control variables for fit (e.g. maxiter)}
}
\description{
Bounded differential expression using count data

Bounded coefficient differential expression
}
