% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit}
\alias{fit}
\title{Fit parameters using the optim function}
\usage{
fit(x, y, weights = rep(1, nrow(x)), offset = 0, control = list(),
  lower = -Inf, upper = Inf)
}
\arguments{
\item{x}{design matrix of dimension ‘n * p’}

\item{y}{vector of observations of length ‘n’}

\item{weights}{vector of weights of length ‘n’}

\item{offset}{vector of coefficient offsets of length ‘n’}

\item{control}{list of control paramters, e.g. ‘maxit’}

\item{lower}{vector of coefficient lower bounds}

\item{upper}{vector of coefficient upper bounds}
}
\value{
named list with estimated coefficients
}
\description{
Fit parameters using the optim function
}
