% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.r
\name{gradient}
\alias{gradient}
\title{Gradient function for negative binomial regression}
\usage{
gradient(par, .x, .y, weights, offset, ...)
}
\arguments{
\item{par}{numeric vector of coefficients in ’x’ and theta}

\item{.x}{design matrix of dimension ‘n * p’}

\item{.y}{vector of observations of length ‘n’}

\item{weights}{vector of weights of length ‘n’}

\item{offset}{vector of coefficient offsets of length ‘n’}

\item{...}{ignored}
}
\value{
direction of steepest descent (the gradient)
}
\description{
The following variables need to be available to the function in order to
be able to call `optim` using this as a gradient function:
}
\details{
x - Regression coefficients matrix
beta - Regressors, such that beta * x = y
y - Observed count values
}
